refactor: 리뷰 작성 가능 캠핑장 조회, 리뷰 작성 API 변경 및 DTO/매퍼 개선 - 서버 로컬 경로 저장 방식의 기존 이미지 업로드 방식의 로직 제거, GCS 사용 방식으로 대체 - 0.5 단위 평점 지원을 위해 reviewRate 타입을 int > double 로 변경(추후 테이블 수정 필요)

+++ workspace_project/Project/src/main/java/com/m4gi/dto/ReviewDTO.java
@@ -13,7 +13,7 @@ public class ReviewDTO {
     private String campgroundId;        // 캠핑장 ID (필터링용)
     private String reservationId;       // 예약 ID (중복 리뷰 체크용)
     private String siteName;			// 사이트명 - 캠핑장 구역용
-    private int reviewRating;           // 별점 점수 (예: 1~5)
+    private double reviewRating;           // 별점 점수 (예: 1~5)
     private String reviewContent;       // 리뷰 내용
     private String reviewPhotosJson;    // 리뷰 사진 URL JSON 문자열
